# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "kasane",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.3.2"
  },
  "paths": {
...
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "ユーザー一覧の取得",
        "description": "ユーザーの一覧を権限つきで取得します。この操作は `global` スコープの `admin` ロールが必要です。",
        "operationId": "list_users",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserInfoResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": [
              "global_admin"
            ]
          }
        ]
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "新規ユーザー作成",
        "description": "新しいユーザーを作成します。`privileges` を指定すると作成と同時に権限を付与できます。\nこの操作は `global` スコープの `admin` ロールが必要です。",
        "operationId": "create_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "権限ルールが不正（global 以外での admin 指定など）"
          },
          "404": {
            "description": "権限ルールが参照するデータベースまたはテーブルが存在しない"
          },
          "409": {
            "description": "同名のユーザーが既に存在する"
          }
        },
        "security": [
          {
            "bearer_auth": [
              "global_admin"
            ]
          }
        ]
      }
    },
    "/users/{username}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "ユーザー情報の取得",
        "description": "指定したユーザーの情報を取得します。本人、または `global` スコープの `admin` ロールを持つ\nユーザーのみ実行可能です。",
        "operationId": "get_user",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "ユーザー名",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example_user"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfoResponse"
                }
              }
            }
          },
          "404": {
            "description": "ユーザーが存在しない"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "ユーザーの削除",
        "description": "指定したユーザーを削除します。この操作は `global` スコープの `admin` ロールが必要です。",
        "operationId": "delete_user",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "ユーザー名",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example_user"
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "403": {
            "description": "rootユーザーは削除不可"
          },
          "404": {
            "description": "ユーザーが存在しない"
          }
        },
        "security": [
          {
            "bearer_auth": [
              "global_admin"
            ]
          }
        ]
      }
    },
    "/users/{username}/password": {
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "パスワードの変更",
        "description": "指定したユーザーのパスワードを変更します。本人、または `global` スコープの `admin` ロールを\n持つユーザーのみ実行可能です。変更すると発行済みトークンはすべて失効します。",
        "operationId": "update_password",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "ユーザー名",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example_user"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": "ユーザーが存在しない"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/users/{username}/privileges": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "権限一覧の取得",
        "description": "指定したユーザーが持つ権限ルールの一覧を取得します。\n\n既に削除されたデータベース・テーブルを指すルールは名前へ解決できないため一覧に現れません\n（そうしたルールは認可判定でも一致しないため、表示と実効権限が食い違うことはありません）。\n\n本人、または `global` スコープの `admin` ロールを持つユーザーのみ実行可能です。",
        "operationId": "get_privileges",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "ユーザー名",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example_user"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivilegesResponse"
                }
              }
            }
          },
          "404": {
            "description": "ユーザーが存在しない"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/users/{username}/privileges/databases/{db_name}": {
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "データベースに対する権限の設定",
        "description": "既に設定されていればロールを置き換えます。配下のテーブルすべてに効きます。\n権限はデータベース ID に紐づくため、改名しても追従します。",
        "operationId": "set_database_privilege",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "ユーザー名",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example_user"
          },
          {
            "name": "db_name",
            "in": "path",
            "description": "データベース名",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example_database"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPrivilegeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "admin はこのスコープに指定できない／上限超過"
          },
          "403": {
            "description": "rootユーザーの権限は変更不可"
          },
          "404": {
            "description": "ユーザーまたはデータベースが存在しない"
          }
        },
        "security": [
          {
            "bearer_auth": [
              "global_admin"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "データベースに対する権限の剥奪",
        "operationId": "delete_database_privilege",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "ユーザー名",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example_user"
          },
          {
            "name": "db_name",
            "in": "path",
            "description": "データベース名",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example_database"
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "403": {
            "description": "rootユーザーの権限は変更不可"
          },
          "404": {
            "description": "ユーザー・データベースが存在しない、または権限を持っていない"
          }
        },
        "security": [
          {
            "bearer_auth": [
              "global_admin"
            ]
          }
        ]
      }
    },
    "/users/{username}/privileges/databases/{db_name}/tables/{table_name}": {
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "テーブルに対する権限の設定",
        "description": "既に設定されていればロールを置き換えます。\n権限はテーブル ID に紐づくため、改名しても追従します。",
        "operationId": "set_table_privilege",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "ユーザー名",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example_user"
          },
          {
            "name": "db_name",
            "in": "path",
            "description": "データベース名",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example_database"
          },
          {
            "name": "table_name",
            "in": "path",
            "description": "テーブル名",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example_table"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPrivilegeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "admin はこのスコープに指定できない／上限超過"
          },
          "403": {
            "description": "rootユーザーの権限は変更不可"
          },
          "404": {
            "description": "ユーザー・データベース・テーブルのいずれかが存在しない"
          }
        },
        "security": [
          {
            "bearer_auth": [
              "global_admin"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "テーブルに対する権限の剥奪",
        "operationId": "delete_table_privilege",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "ユーザー名",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example_user"
          },
          {
            "name": "db_name",
            "in": "path",
            "description": "データベース名",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example_database"
          },
          {
            "name": "table_name",
            "in": "path",
            "description": "テーブル名",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example_table"
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "403": {
            "description": "rootユーザーの権限は変更不可"
          },
          "404": {
            "description": "対象が存在しない、または権限を持っていない"
          }
        },
        "security": [
          {
            "bearer_auth": [
              "global_admin"
            ]
          }
        ]
      }
    },
    "/users/{username}/privileges/global": {
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "サーバー全体に対する権限の設定",
        "description": "既に設定されていればロールを置き換えます。`admin` を指定できるのはこのスコープだけです。",
        "operationId": "set_global_privilege",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "ユーザー名",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example_user"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPrivilegeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "保持できる権限数の上限を超えた"
          },
          "403": {
            "description": "rootユーザーの権限は変更不可"
          },
          "404": {
            "description": "ユーザーが存在しない"
          }
        },
        "security": [
          {
            "bearer_auth": [
              "global_admin"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "サーバー全体に対する権限の剥奪",
        "description": "ロールを問わず、このスコープの権限を取り除きます。",
        "operationId": "delete_global_privilege",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "ユーザー名",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example_user"
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "403": {
            "description": "rootユーザーの権限は変更不可"
          },
          "404": {
            "description": "ユーザーが存在しない、またはこのスコープの権限を持っていない"
          }
        },
        "security": [
          {
            "bearer_auth": [
              "global_admin"
            ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateUserRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string",
            "example": "example_user"
          },
          "password": {
            "type": "string",
            "example": "secret123"
          },
          "privileges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PrivilegeRule"
            },
            "description": "作成と同時に付与する権限。省略時は権限なし。"
          }
        }
      },
...
      "PrivilegeRule": {
        "oneOf": [
          {
            "type": "object",
            "title": "GlobalPrivilege",
            "description": "サーバー全体。`role` に `admin` を指定できるのはこのスコープだけ。",
            "required": [
              "role",
              "scope"
            ],
            "properties": {
              "role": {
                "$ref": "#/components/schemas/UserRole"
              },
              "scope": {
                "type": "string",
                "enum": [
                  "global"
                ]
              }
            }
          },
          {
            "type": "object",
            "title": "DatabasePrivilege",
            "description": "データベース配下すべて。",
            "required": [
              "db_name",
              "role",
              "scope"
            ],
            "properties": {
              "db_name": {
                "type": "string",
                "example": "example_database"
              },
              "role": {
                "$ref": "#/components/schemas/UserRole"
              },
              "scope": {
                "type": "string",
                "enum": [
                  "database"
                ]
              }
            }
          },
          {
            "type": "object",
            "title": "TablePrivilege",
            "description": "単一テーブル。",
            "required": [
              "db_name",
              "table_name",
              "role",
              "scope"
            ],
            "properties": {
              "db_name": {
                "type": "string",
                "example": "example_database"
              },
              "table_name": {
                "type": "string",
                "example": "example_table"
              },
              "role": {
                "$ref": "#/components/schemas/UserRole"
              },
              "scope": {
                "type": "string",
                "enum": [
                  "table"
                ]
              }
            }
          }
        ],
        "description": "API 上の権限ルール表現。\n\n保存形式（[`StoredPrivilege`](super::entity::StoredPrivilege)）は UUID ベースだが、\n外部にはデータベース名・テーブル名で見せる。付与時に名前 → ID を解決し、\n存在しない名前は 404 で弾く。"
      },
      "PrivilegesResponse": {
        "type": "object",
        "description": "`GET /users/{username}/privileges` のレスポンス。\n\n個々のルールの追加・変更・削除は、対象ごとのサブリソース\n（`.../privileges/global`、`.../privileges/databases/{db_name}` など）へ\n`PUT` / `DELETE` する。",
        "required": [
          "privileges"
        ],
        "properties": {
          "privileges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PrivilegeRule"
            }
          }
        }
      },
...
      "SetPrivilegeRequest": {
        "type": "object",
        "description": "1 つの対象に対する権限の設定リクエスト。対象はパスで指定する。",
        "required": [
          "role"
        ],
        "properties": {
          "role": {
            "$ref": "#/components/schemas/UserRole"
          }
        }
      },
...
      "UpdatePasswordRequest": {
        "type": "object",
        "required": [
          "password"
        ],
        "properties": {
          "password": {
            "type": "string",
            "example": "new_secret123"
          }
        }
      },
...
      "UserInfoResponse": {
        "type": "object",
        "required": [
          "username",
          "privileges"
        ],
        "properties": {
          "username": {
            "type": "string",
            "example": "example_user"
          },
          "privileges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PrivilegeRule"
            },
            "description": "現在有効な権限ルール。既に削除されたデータベース／テーブルを指すルールは\n解決できないため一覧に現れない（実際の認可でも一致しない）。"
          }
        }
      },
      "UserRole": {
        "type": "string",
        "description": "権限の強さ。数値が大きいほど強く、上位は下位をすべて含む。\n\n`admin` だけはデータ面ではなく制御面（ユーザーの作成・削除・権限付与）を表し、\n`global` スコープにのみ指定できる。`database` / `table` スコープで `admin` を\n指定したリクエストは 400 で拒否される。",
        "enum": [
          "read",
          "write",
          "manage",
          "admin"
        ]
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}